Time: 1789 ms (12.50%), Space: 148.9 MB (62.50%) - LeetHub

+++ 2416-sum-of-prefix-scores-of-strings/2416-sum-of-prefix-scores-of-strings.cpp
@@ -22,7 +22,7 @@ public:
             long long int p = 29;
             int c = 0;
             for(int i = 0 ; i < it.size() ; i++){
-                hash = (hash + (long long)(it[i] - 'a' + 1)*po) % m;
+                hash = (hash + (it[i] - 'a' + 1)*po) % m;
                 po = (po * p) % m;
